+++ src/libraries/UtilsLib.sol
@@ -68,17 +68,22 @@ library UtilsLib {
         }
     }
 
-    function msb(uint256 bitmap) internal pure returns (uint256 res) {
+    /// @dev Assumes `bitmap` is not zero.
+    function msb(uint128 bitmap) internal pure returns (uint256 res) {
         assembly {
             res := sub(255, clz(bitmap))
         }
     }
 
-    function setBit(uint256 bitmap, uint256 bit) internal pure returns (uint256) {
-        return bitmap | (1 << bit);
+    /// @dev Assumes `bit` is less than 128.
+    function setBit(uint128 bitmap, uint256 bit) internal pure returns (uint128) {
+        // forge-lint: disable-next-item(unsafe-typecast) as bit < 128
+        return uint128(bitmap | (1 << bit));
     }
 
-    function clearBit(uint256 bitmap, uint256 bit) internal pure returns (uint256) {
-        return bitmap & ~(1 << bit);
+    /// @dev Assumes `bit` is less than 128.
+    function clearBit(uint128 bitmap, uint256 bit) internal pure returns (uint128) {
+        // forge-lint: disable-next-item(unsafe-typecast)
+        return uint128(bitmap & ~(1 << bit));
     }
 }
